supervisor/requirements.txt
# Supervisor dependencies
fastapi>=0.115.5
uvicorn>=0.34.0
websockets>=13.0
anthropic[bedrock]>=0.41.0
boto3>=1.35.90
openai>=1.54.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
pytest>=8.3.3
